Cleaned up worker table a bit.
