+++ ui/.env.development
@@ -1,0 +1,1 @@
+API_SERVER_URL=http://api:3000/
